+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Page({ params }: { params: { projectId: string } }) {
-  redirect(`/projects/${params.projectId}/users`);
-}