+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @stackframe/stack-backend
 
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+  - @stackframe/stack-emails@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack-emails@2.7.2
+
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-emails@2.7.1
+  - @stackframe/stack-shared@2.7.1
+
 ## 2.7.0
 
 ### Minor Changes