+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @stackframe/docs-examples
 
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+  - @stackframe/stack@2.7.3
+  - @stackframe/stack-ui@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack@2.7.2
+  - @stackframe/stack-ui@2.7.2
+
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.1
+  - @stackframe/stack-ui@2.7.1
+  - @stackframe/stack@2.7.1
+
 ## 2.7.0
 
 ### Patch Changes