+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @stackframe/stack-shared
 
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.2
+
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.1
+
 ## 2.7.0
 
 ### Minor Changes