+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @stackframe/stack
 
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+  - @stackframe/stack-ui@2.7.3
+  - @stackframe/stack-sc@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack-ui@2.7.2
+  - @stackframe/stack-sc@2.7.2
+
+## 2.7.1
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.1
+  - @stackframe/stack-ui@2.7.1
+  - @stackframe/stack-sc@2.7.1
+
 ## 2.7.0
 
 ### Minor Changes